+++ src/truepyfstlib/__init__.py
@@ -25,7 +25,7 @@ from .common import (
     FstHierType,
 )
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 __all__ = [
     "FstReader",
     "FstWriter",